(BOLT-608) Add `local` transport for Windows OS
Add support for local transport on windows OS. The WinRM and Local transports now share some common powershell code generation functionality. The powershell templating has been pulled out into a `Powerhsell` module in order to share.
