+++ lib/bolt/transport/winrm.rb
@@ -2,14 +2,11 @@
 
 require 'bolt/node/errors'
 require 'bolt/transport/base'
+require 'bolt/transport/powershell'
 
 module Bolt
   module Transport
     class WinRM < Base
-      PS_ARGS = %w[
-        -NoProfile -NonInteractive -NoLogo -ExecutionPolicy Bypass
-      ].freeze
-
       def self.options
         %w[port user password connect-timeout ssl ssl-verify tmpdir cacert extensions]
       end
@@ -78,34 +75,14 @@ module Bolt
       def run_script(target, script, arguments, _options = {})
         # unpack any Sensitive data
         arguments = unwrap_sensitive_args(arguments)
-
         with_connection(target) do |conn|
           conn.with_remote_tempdir do |dir|
             remote_path = conn.write_remote_executable(dir, script)
-            if powershell_file?(remote_path)
-              mapped_args = arguments.map do |a|
-                "$invokeArgs.ArgumentList += @'\n#{a}\n'@"
-              end.join("\n")
-              output = conn.execute(<<-PS)
-$invokeArgs = @{
-  ScriptBlock = (Get-Command "#{remote_path}").ScriptBlock
-  ArgumentList = @()
-}
-#{mapped_args}
-
-try
-{
-  Invoke-Command @invokeArgs
-}
-catch
-{
-  Write-Error $_.Exception
-  exit 1
-}
-          PS
+            if Powershell.powershell_file?(remote_path)
+              output = conn.execute(Powershell.run_script(arguments, remote_path))
             else
-              path, args = *process_from_extension(remote_path)
-              args += escape_arguments(arguments)
+              path, args = *Powershell.process_from_extension(remote_path)
+              args += Powershell.escape_arguments(arguments)
               output = conn.execute_process(path, args)
             end
             Bolt::Result.for_command(target, output.stdout.string, output.stderr.string, output.exit_code)
@@ -118,7 +95,7 @@ catch
         executable = implementation['path']
         input_method = implementation['input_method']
         extra_files = implementation['files']
-        input_method ||= powershell_file?(executable) ? 'powershell' : 'both'
+        input_method ||= Powershell.powershell_file?(executable) ? 'powershell' : 'both'
 
         # unpack any Sensitive data
         arguments = unwrap_sensitive_args(arguments)
@@ -144,7 +121,7 @@ catch
 
             if ENVIRONMENT_METHODS.include?(input_method)
               envify_params(arguments).each do |(arg, val)|
-                cmd = "[Environment]::SetEnvironmentVariable('#{arg}', @'\n#{val}\n'@)"
+                cmd = Powershell.set_env(arg, val)
                 result = conn.execute(cmd)
                 if result.exit_code != 0
                   raise Bolt::Node::EnvironmentVarError.new(arg, val)
@@ -154,71 +131,16 @@ catch
 
             conn.shell_init
             output =
-              if powershell_file?(remote_task_path) && stdin.nil?
-                # NOTE: cannot redirect STDIN to a .ps1 script inside of PowerShell
-                # must create new powershell.exe process like other interpreters
-                # fortunately, using PS with stdin input_method should never happen
-                if input_method == 'powershell'
-                  conn.execute(<<-PS)
-$private:tempArgs = Get-ContentAsJson (
-  $utf8.GetString([System.Convert]::FromBase64String('#{Base64.encode64(JSON.dump(arguments))}'))
-)
-$allowedArgs = (Get-Command "#{remote_task_path}").Parameters.Keys
-$private:taskArgs = @{}
-$private:tempArgs.Keys | ? { $allowedArgs -contains $_ } | % { $private:taskArgs[$_] = $private:tempArgs[$_] }
-try { & "#{remote_task_path}" @taskArgs } catch { Write-Error $_.Exception; exit 1 }
-              PS
-                else
-                  conn.execute(%(try { & "#{remote_task_path}" } catch { Write-Error $_.Exception; exit 1 }))
-                end
+              if Powershell.powershell_file?(remote_task_path) && stdin.nil?
+                conn.execute(Powershell.run_ps_task(arguments, remote_task_path, input_method))
               else
-                path, args = *process_from_extension(remote_task_path)
+                path, args = *Powershell.process_from_extension(remote_task_path)
                 conn.execute_process(path, args, stdin)
               end
 
             Bolt::Result.for_task(target, output.stdout.string,
                                   output.stderr.string,
                                   output.exit_code)
-          end
-        end
-      end
-
-      def powershell_file?(path)
-        Pathname(path).extname.casecmp('.ps1').zero?
-      end
-
-      def process_from_extension(path)
-        case Pathname(path).extname.downcase
-        when '.rb'
-          [
-            'ruby.exe',
-            ['-S', "\"#{path}\""]
-          ]
-        when '.ps1'
-          [
-            'powershell.exe',
-            [*PS_ARGS, '-File', "\"#{path}\""]
-          ]
-        when '.pp'
-          [
-            'puppet.bat',
-            ['apply', "\"#{path}\""]
-          ]
-        else
-          # Run the script via cmd, letting Windows extension handling determine how
-          [
-            'cmd.exe',
-            ['/c', "\"#{path}\""]
-          ]
-        end
-      end
-
-      def escape_arguments(arguments)
-        arguments.map do |arg|
-          if arg =~ / /
-            "\"#{arg}\""
-          else
-            arg
           end
         end
       end
