+++ spec/integration/local_spec.rb
@@ -4,31 +4,45 @@ require 'spec_helper'
 require 'bolt_spec/files'
 require 'bolt_spec/integration'
 
-describe "when running over the local transport", bash: true do
+describe "when running over the local transport" do
   include BoltSpec::Files
   include BoltSpec::Integration
 
-  let(:whoami) { "whoami" }
   let(:modulepath) { File.join(__dir__, '../fixtures/modules') }
-  let(:stdin_task) { "sample::stdin" }
   let(:uri) { 'localhost,local://foo' }
   let(:user) { ENV['USER'] }
 
   after(:each) { Puppet.settings.send(:clear_everything_for_tests) }
 
   context 'when using CLI options' do
-    let(:config_flags) {
-      %W[--nodes #{uri} --no-host-key-check --format json --modulepath #{modulepath}]
-    }
+    let(:echo) { "echo hi" }
 
     it 'runs multiple commands' do
-      result = run_nodes(%W[command run #{whoami}] + config_flags)
-      expect(result.map { |r| r['stdout'].strip }).to eq([user, user])
+      result = run_nodes(%W[command run #{echo} --nodes #{uri} --format json])
+      expect(result.map { |r| r['stdout'].strip }).to eq(%w[hi hi])
     end
 
     it 'reports errors when command fails' do
-      result = run_failed_nodes(%w[command run boop] + config_flags)
+      result = run_failed_nodes(%W[command run boop --nodes #{uri} --format json])
       expect(result[0]['_error']).to be
+    end
+  end
+
+  context 'when using CLI options on POSIX OS', bash: true do
+    let(:config_flags) {
+      %W[--nodes #{uri} --no-host-key-check --format json --modulepath #{modulepath}]
+    }
+    let(:whoami) { "whoami" }
+    let(:stdin_task) { "sample::stdin" }
+
+    it 'runs script with parameter', :reset_puppet_settings do
+      with_tempfile_containing('script', "#!/usr/bin/env bash \n echo $1", '.sh') do |script|
+        results = run_cli_json(%W[script run #{script.path} param --nodes localhost])
+        results['items'].each do |result|
+          expect(result['status']).to eq('success')
+          expect(result['result']).to eq("stdout" => "param\n", "stderr" => "", "exit_code" => 0)
+        end
+      end
     end
 
     it 'runs multiple tasks', :reset_puppet_settings do
@@ -39,6 +53,65 @@ describe "when running over the local transport", bash: true do
     it 'reports errors when task fails', :reset_puppet_settings do
       result = run_failed_nodes(%w[task run results fail=true] + config_flags)
       expect(result[0]['_error']).to be
+    end
+  end
+
+  context 'when using CLI options on Windows OS', windows: true do
+    let(:config_flags) {
+      %W[--nodes localhost --format json --modulepath #{modulepath}]
+    }
+
+    it 'runs powershell script with parameter', :reset_puppet_settings do
+      with_tempfile_containing('script', "Write-Host $args", '.ps1') do |script|
+        results = run_cli_json(%W[script run #{script.path} param -n localhost])
+        results['items'].each do |result|
+          expect(result['status']).to eq('success')
+          expect(result['result']).to eq("stdout" => "param\n", "stderr" => "", "exit_code" => 0)
+        end
+      end
+    end
+
+    it 'runs ruby script with parameter', :reset_puppet_settings do
+      ruby_script = "puts 'Ruby' \n ARGV.each {|a| puts a}"
+      with_tempfile_containing('script', ruby_script, '.rb') do |script|
+        results = run_cli_json(%W[script run #{script.path} param -n localhost])
+        results['items'].each do |result|
+          expect(result['status']).to eq('success')
+          expect(result['result']).to eq("stdout" => "Ruby\nparam\n", "stderr" => "", "exit_code" => 0)
+        end
+      end
+    end
+
+    it 'runs a task reading from stdin', :reset_puppet_settings do
+      result = run_one_node(%w[task run sample::winstdin message=somemessage] + config_flags)
+      output = result['_output'].strip
+      expect(output).to match(/STDIN: {"message":"somemessage"/)
+    end
+
+    it 'runs a task reading from $input', :reset_puppet_settings do
+      result = run_one_node(%w[task run sample::wininput message=somemessage] + config_flags)
+      output = result['_output'].strip
+      expect(output).to match(/INPUT: {"message":"somemessage"/)
+    end
+
+    it 'runs a task with parameters', :reset_puppet_settings do
+      result = run_one_node(%w[task run sample::winparams message=µsomemessage] + config_flags)
+      output = result['_output'].strip
+      expect(output).to match(/Message: µsomemessage/)
+    end
+
+    it 'runs a task reading from environment variables', :reset_puppet_settings do
+      result = run_one_node(%w[task run sample::winenv message=µsomemessage] + config_flags)
+      output = result['_output'].strip
+      expect(output).to match(/ENV: µsomemessage/)
+    end
+
+    it 'runs a task with complex parameters', :reset_puppet_settings do
+      complex_input_file = File.join(__dir__, '../fixtures/complex_params/input.json')
+      expected = File.open(File.join(__dir__, '../fixtures/complex_params/output'), 'rb', &:read)
+      expected = expected.gsub(/\r\n?/, "\n")
+      result = run_one_node(%W[task run sample::complex_params --params @#{complex_input_file}] + config_flags)
+      expect(result['_output']).to eq(expected)
     end
   end
 end
